Assignment 2: ASP.Net Core DI and IoC
Design a C# console application that demonstrates the implementation of Dependency
Injection using the built-in .NET Core DI container. The application should have a simple
class hierarchy with an interface and two implementations of that interface.
1. Create an interface called `ICalculator` with at least two methods: `int Add(int a, int
b)` and `int Subtract(int a, int b)`.
2. Implement the `ICalculator` interface in two different classes: `SimpleCalculator` and
`AdvancedCalculator`. The `SimpleCalculator` should implement the `Add` and
`Subtract` methods by using the basic `+` and `-` operators. The
`AdvancedCalculator` should implement the `Add` and `Subtract` methods by using
more advanced mathematical algorithms.
3. Create a class called `CalculatorController` that has two private fields: an instance of
`ICalculator` and a string called `CalculatorType`. The `CalculatorController` should
have a constructor that takes an `ICalculator` instance and a string representing the
type of calculator being used. The `CalculatorController` should also have a public
method called `Calculate` that takes two integers and either adds or subtracts them
depending on the value of `CalculatorType`.
4. Use the built-in .NET Core DI container to inject one of the implementations of
`ICalculator` into the constructor of the `CalculatorController` class.
5. In the `Main` method, create an instance of the `CalculatorController` class and call
the `Calculate` method with two integers.